Changes to have grammar2.json to process correctly.
Whenever a repeat keyword is found, a new intermediate node and a new
statement are created. The new intermediate node is expanded as
List(IntermediateX) and the new statement is added to the list of new
statements that is then integrated with the original program list. When
the new statement is added to the program statement list, it is inserted
immediately after its use as in the grammar.json example